fix: dont delete bot messages

+++ src/badtranslator.ts
@@ -24,9 +24,8 @@ export default class BadTranslator {
 
     async init() {
         this.bot.client.on('messageCreate', async ({message}) => {
-            if (!this.channels.has(message.channelId)) return;
+            if (!this.channels.has(message.channelId) || message.author.bot) return;
             if (message.content.length === 0 || message.content.length > MAX_MESSAGE_LENGTH) return message.delete();
-            if (message.author.bot) return message.delete();
 
             const isRatelimited = this.isRatelimited(message.author.id);
             if (isRatelimited) {
